Fire a Perl callback on probe trigger
This introduces Devel::Probe::trigger, a setter for the Perl sub which
should fire when a probe triggers. The sub is invoked in void context.

This allows tricks like using `caller`, Devel::StackTrace, or PadWalker
`peek_my` to inspect the environment and export the resulting data
somewhere.

Hat tip to perlxstut and perlcall.