Move config parsing to Perl, cleanup code
Much easier to parse the config file from Perl.  This required exposing
some more stuff from the C implementation, to be called from the Perl
parsing routine.  But that is fine, this also makes it easier to test
and benchmark the code (future work).

Cleaned up a lot of cruft too.