Simplify API
We now expose more building blocks so that you can import the module,
configure and enable it at your heart's content; see the perldoc for
more details.

This also means that several tests were out of date, and we can now
fully test the module in t/007-trigger.t.  Yay.