Test for install/reinstall.
Runtime enable/disable keeps overhead low when probing is not being
used: Devel::Probe in disabled state introduced an additional ~500
micros mean, and ~1-2ms at p99.

However, in a perfect world we'd totally unhook the OP_NEXTSTATE when
probing was not being used. This would reduce the overhead to 0.

This test checks for our ability to dynamically install/remove/install
during the process lifetime and have the expected probes fire.