Support a custom argument to probe trigger.
This patch introduces an optional custom argument for probe callbacks.
The argument is a single arbitrary Perl scalar. That means you can do
the following:

    Devel::Probe::trigger(sub {
        my ($file, $line, $mood) = @_;
        say "$file/$line is $mood";
    });

    Devel::Probe::add_probe("foo.pl", 42, Devel::Probe::ONCE, "so jazzed");
    Devel::Probe::add_probe("foo.pl", 12, Devel::Probe::ONCE, "fine, I guess");

This enables very powerful probes of the following sort:

    use PadWalker qw(peek_my);
    Devel::Probe::trigger(sub {
        my ($file, $line, $vars) = @_;
        my $scope = peek_my(1);
        for my $var (@$vars) {
            say $scope->{$var} // "not defined";
        }
    });

    Devel::Probe::add_probe("foo.pl", 42, Devel::Probe::ONCE, ['$interesting_var']);
    # some lines later ...
    my $interesting_var = "fascinating";

This was done by changing the probe_hash data structure from this:

    {
	"foo.pl" => {
	    10 => 1,
	    15 => 2,
	}
    }

To this:

    {
        "foo.pl" => {
            10 => [1, { user => 'btyler', vars => [ '$user_id', '$log_level' ] },
            15 => [2, "halt-and-catch-fire" ],
        }
    }

The first index in the AV is the probe type, the second is the user's
argument (if any). If no argument was defined in add_probe, the AV will
only have one item.

Some implementation choices:

1) probe_lookup no longer returns probe flags, just 'exists'.

This was motivated by not wanting to introduce more complexity to
probe_lookup, which is the super hot path. I think the split into
'probe_lookup' and 'probe_settings' is pretty reasonable because the
perf criticality of probe_lookup (every line of Perl code) is
meaningfully different from probe_settings (only once we know we want to
fire a probe).

2) Custom probe callback argument is optional, and your callback will
only get 2 parameters (file/line) if you don't specify an arg.

I did a bit of mucking around here because I didn't want to
unconditionally store 'undef' in the second position of the AV
containing a probe's settings, or even weirder, pass it to the probe
callback.